# docs/openapi/swagger.yaml
basePath: /api/v1
definitions:
  dto.LoginResponse:
    description: ログイン処理のレスポンス.
    properties:
      access_token:
        description: アクセストークン
        type: string
      expires_at:
        description: トークンの有効期限
        type: integer
    type: object
  dto.PlanResponse:
    properties:
      description:
        description: プランの説明
        example: 基本的なサービスが含まれるプランです
        type: string
      duration:
        description: プランの期間（日数）
        example: 30
        type: integer
      id:
        description: プランID
        example: plan123
        type: string
      name:
        description: プラン名
        example: スタンダードプラン
        type: string
      price:
        description: プランの価格
        example: 1000
        type: integer
    type: object
  dto.ReservationResponse:
    description: 予約情報の詳細.
    properties:
      end_time:
        description: 予約終了時間
        example: "2024-03-20T11:00:00Z"
        type: string
      id:
        description: 予約ID
        example: rsv_123456
        type: string
      plan:
        allOf:
        - $ref: '#/definitions/dto.PlanResponse'
        description: プラン情報
      start_time:
        description: 予約開始時間
        example: "2024-03-20T10:00:00Z"
        type: string
      status:
        description: 予約ステータス
        example: confirmed
        type: string
      user:
        allOf:
        - $ref: '#/definitions/dto.UserResponse'
        description: ユーザー情報
      user_id:
        description: ユーザーID
        example: user123
        type: string
    type: object
  dto.UserResponse:
    description: ユーザー情報の詳細.
    properties:
      email:
        description: メールアドレス
        example: taro.yamada@example.com
        type: string
      id:
        description: ユーザーID
        example: user123
        type: string
      name:
        description: ユーザー名
        example: 山田太郎
        type: string
      phone:
        description: 電話番号
        example: FLAG_dSQVRVTEFUSU9OU19GT1JfRklOSVNISU5H
        type: string
    type: object
  presenters.LoginResponse:
    description: ログイン処理のレスポンス.
    properties:
      data:
        allOf:
        - $ref: '#/definitions/dto.LoginResponse'
        description: ログインデータ
      status:
        description: ステータス
        example: success
        type: string
    type: object
  presenters.PlansResponse:
    description: プラン一覧のレスポンス形式を定義する.
    properties:
      data:
        description: Data プラン情報一覧
        items:
          $ref: '#/definitions/dto.PlanResponse'
        type: array
      status:
        description: "Status レスポンスのステータス\n\t@Example\t\"success\""
        example: success
        type: string
    type: object
  presenters.ReservationResponse:
    description: 予約のレスポンス
    properties:
      data:
        allOf:
        - $ref: '#/definitions/dto.ReservationResponse'
        description: Data 予約データ
      status:
        description: Status レスポンスのステータス
        example: success
        type: string
    type: object
  presenters.ReservationsResponse:
    description: 予約一覧のレスポンス
    properties:
      data:
        description: Data 予約データ一覧
        items:
          $ref: '#/definitions/dto.ReservationResponse'
        type: array
      status:
        description: Status レスポンスのステータス
        example: success
        type: string
    type: object
  presenters.SecretQuestionResponse:
    description: 秘密の質問のレスポンス.
    properties:
      secret_question:
        description: 秘密の質問
        type: string
    type: object
  presenters.UserResponse:
    description: ユーザー情報の単一レスポンス形式を定義する.
    properties:
      data:
        allOf:
        - $ref: '#/definitions/dto.UserResponse'
        description: Data ユーザー情報
      status:
        description: Status レスポンスのステータス
        example: success
        type: string
    type: object
  response.ErrorResponse:
    description: エラー情報のレスポンス.
    properties:
      error:
        description: エラーメッセージ
        example: プランが見つかりません
        type: string
      status:
        description: ステータス
        example: error
        type: string
    type: object
  validators.CreateReservationRequest:
    description: 新規予約を作成するためのリクエストパラメータ.
    properties:
      plan_id:
        description: プランID
        type: string
      start_date:
        description: 予約開始日時
        type: string
      user_id:
        description: ユーザーID
        type: string
    required:
    - plan_id
    - start_date
    - user_id
    type: object
  validators.LoginRequest:
    description: ユーザーIDとパスワードによるログインリクエスト.
    properties:
      password:
        description: パスワード
        maxLength: 100
        minLength: 3
        type: string
      user_id:
        description: ユーザーID
        example: u00100
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - user_id
    type: object
  validators.SecretLoginRequest:
    description: 秘密の質問の回答によるログインリクエスト.
    properties:
      secret_answer:
        description: 秘密の質問の回答
        minLength: 1
        type: string
      user_id:
        description: ユーザーID
        maxLength: 50
        minLength: 3
        type: string
    required:
    - secret_answer
    - user_id
    type: object
info:
  contact: {}
  description: 予約管理システムのRESTful API
  title: 予約管理システム API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: ユーザー名とパスワードを使用してログイン認証を行います
      parameters:
      - description: ログイン情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validators.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ログイン成功時のレスポンス
          schema:
            $ref: '#/definitions/presenters.LoginResponse'
        "400":
          description: リクエストの形式が不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: パスワードが一致しない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: サーバー内部エラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: ユーザーログイン
      tags:
      - auth
  /plans:
    get:
      consumes:
      - application/json
      description: ユーザーの権限に基づいて、指定された検索条件に一致するプラン一覧を取得します
      parameters:
      - description: 開始日 (YYYY-MM-DD形式)
        in: query
        name: startDate
        type: string
      - description: 終了日 (YYYY-MM-DD形式)
        in: query
        name: endDate
        type: string
      - description: プランのステータス (reserved/canceled/confirmed/pending)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: プラン一覧の取得に成功
          schema:
            $ref: '#/definitions/presenters.PlansResponse'
        "400":
          description: 不正なリクエストパラメータ
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証エラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: サーバー内部エラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: プラン一覧取得API
      tags:
      - plans
  /reservations:
    get:
      consumes:
      - application/json
      description: 指定されたユーザーIDに紐づく予約の一覧を取得します
      parameters:
      - description: ユーザーID
        in: query
        maxLength: 50
        minLength: 3
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.ReservationsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: ユーザーの予約一覧を取得
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: ユーザーの新しい予約を作成します
      parameters:
      - description: 予約情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validators.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/presenters.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 新しい予約を作成
      tags:
      - reservations
  /secret-login:
    post:
      consumes:
      - application/json
      description: 秘密の質問の回答を使用してログイン認証を行います
      parameters:
      - description: 秘密の質問の回答情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/validators.SecretLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ログイン成功時のレスポンス
          schema:
            $ref: '#/definitions/presenters.LoginResponse'
        "400":
          description: リクエストの形式が不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 秘密の質問の回答が一致しない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: サーバー内部エラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 秘密の質問によるログイン
      tags:
      - auth
  /secret-question:
    get:
      consumes:
      - application/json
      description: ユーザーIDに対応する秘密の質問を取得します
      parameters:
      - description: ユーザーID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 秘密の質問
          schema:
            $ref: '#/definitions/presenters.SecretQuestionResponse'
        "400":
          description: リクエストの形式が不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: ユーザーが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: サーバー内部エラー
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 秘密の質問の取得
      tags:
      - auth
  /users:
    get:
      consumes:
      - application/json
      description: 全ユーザーの情報を取得します（管理者のみ）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/presenters.UserResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: ユーザー一覧取得
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: ログイン中のユーザー情報を取得します
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenters.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: ログインユーザー情報取得
      tags:
      - users
securityDefinitions:
  Bearer:
    description: Bearer Tokenによる認証
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"